# Improve flag and string widget performance