--- a/src/widgets/FlagsWidget.cpp
+++ b/src/widgets/FlagsWidget.cpp
@@ -8,6 +8,7 @@
 #include <QMenu>
 #include <QShortcut>
 #include <QTreeWidget>
+#include <QStandardItemModel>
 
 FlagsModel::FlagsModel(QList<FlagDescription> *flags, QObject *parent)
     : AddressableItemModel<QAbstractListModel>(parent),
@@ -82,6 +83,14 @@ QString FlagsModel::name(const QModelIndex &index) const
     return flag.name;
 }
 
+const FlagDescription *FlagsModel::description(QModelIndex index) const
+{
+    if (index.row() < flags->size()) {
+        return &flags->at(index.row());
+    }
+    return nullptr;
+}
+
 FlagsSortFilterProxyModel::FlagsSortFilterProxyModel(FlagsModel *source_model, QObject *parent)
     : AddressableFilterProxyModel(source_model, parent)
 {
@@ -96,26 +105,27 @@ bool FlagsSortFilterProxyModel::filterAcceptsRow(int row, const QModelIndex &par
 
 bool FlagsSortFilterProxyModel::lessThan(const QModelIndex &left, const QModelIndex &right) const
 {
-    FlagDescription left_flag = left.data(FlagsModel::FlagDescriptionRole).value<FlagDescription>();
-    FlagDescription right_flag = right.data(FlagsModel::FlagDescriptionRole).value<FlagDescription>();
+    auto source = static_cast<FlagsModel*>(sourceModel());
+    auto left_flag = source->description(left);
+    auto right_flag = source->description(right);
 
     switch (left.column()) {
     case FlagsModel::SIZE:
-        if (left_flag.size != right_flag.size)
-            return left_flag.size < right_flag.size;
+        if (left_flag->size != right_flag->size)
+            return left_flag->size < right_flag->size;
     // fallthrough
     case FlagsModel::OFFSET:
-        if (left_flag.offset != right_flag.offset)
-            return left_flag.offset < right_flag.offset;
+        if (left_flag->offset != right_flag->offset)
+            return left_flag->offset < right_flag->offset;
     // fallthrough
     case FlagsModel::NAME:
-        return left_flag.name < right_flag.name;
+        return left_flag->name < right_flag->name;
     default:
         break;
     }
 
     // fallback
-    return left_flag.offset < right_flag.offset;
+    return left_flag->offset < right_flag->offset;
 }
 
 
@@ -212,6 +222,7 @@ void FlagsWidget::refreshFlagspaces()
     if (cur_idx < 0)
         cur_idx = 0;
 
+    disableFlagRefresh = true; // prevent duplicate flag refresh caused by flagspaceCombo modifications
     ui->flagspaceCombo->clear();
     ui->flagspaceCombo->addItem(tr("(all)"));
 
@@ -221,12 +232,16 @@ void FlagsWidget::refreshFlagspaces()
 
     if (cur_idx > 0)
         ui->flagspaceCombo->setCurrentIndex(cur_idx);
+    disableFlagRefresh = false;
 
     refreshFlags();
 }
 
 void FlagsWidget::refreshFlags()
 {
+    if (disableFlagRefresh) {
+        return;
+    }
     QString flagspace;
 
     QVariant flagspace_data = ui->flagspaceCombo->currentData();
@@ -238,7 +253,7 @@ void FlagsWidget::refreshFlags()
     flags = Core()->getAllFlags(flagspace);
     flags_model->endResetModel();
 
-    qhelpers::adjustColumns(ui->flagsTreeView, 2, 0);
+    //qhelpers::adjustColumns(ui->flagsTreeView, 2, 0);
 
     tree->showItemsNumber(flags_proxy_model->rowCount());