--- a/src/widgets/StringsWidget.h
+++ b/src/widgets/StringsWidget.h
@@ -42,6 +42,7 @@ public:
     QVariant headerData(int section, Qt::Orientation orientation, int role = Qt::DisplayRole) const override;
 
     RVA address(const QModelIndex &index) const override;
+    const StringDescription *description(const QModelIndex &index) const;
 };