<!-- Any change needs to be discussed before proceeding. Failure to do so may result in the rejection of the pull request. -->


**Detailed description**

Times bellow are from test executable described in #2020 

* Use refreshDeferrer for VtableWidget saves ~4s if its not used
* Don't do duplicate flag refreshes 3x450ms -> 1x450ms
* setUniformRowHeights(false) - saves ~ 4-5s for strings widget and flag widget each . 
* replace qHelpers::adjustColumns with appropriate Qt properties. String widget now stretches the string column instead of last one containing section.  adjustColumns might not be as bad as I thought, it might have triggered the rowHeight recalculation causing misleading time measurements.
* reduce the overhead during sorting, saves a few hundred ms in debug mode. Didn't retest in release mode.
* Don't sort strings unless requested by user, order returned by r2 is somewhat reasonable.

**Test plan (required)**

* Open test executable described in #2020 and strings widget
* Press F5 to refresh all, observe delay. Took 25s in debug and 21s in release mode before fix, 4.7s after the fix.
* try searching string "tst" quick filter and changing sorting column. Took 3-4s before fix, ~100ms after even when pattern matches most of the strings.  Different search queries like "123" may take longer, but it is still better than before.
* Searching in flag widget should have similar improvements as strings.

<!-- **Code formatting**
Make sure you ran astyle on your code before making the PR. Check our contribution guidelines here: https://cutter.re/docs/code.html -->

**Other issues observed during testing**
* Sections selection in strings widget are broken, also affects master
* Trying to debug the test executable causes some kind of deadlock. Repeatable using master without this PR. Rebase signal triggers widget refresh, which causes access to r2 and everything gets stuck
* Section listing always triggers reading of memory for calculating checksums even when list of requested checksums is empty due to `"" != NULL`. Might cause unnecessary slowdown especially during debugging. 

**Closing issues**

First part for #2020 doesn't close it. Other widgets need to be tested with appropriate test executables.
<!-- put `closes #XXXX` in your comment to auto-close the issue that your PR fixes (if such). -->