@ITAYC0HEN 2020-01-20
Much better. Noticeable improvements, well done!
sorting and searching are indeed much better.

**observations:**
 - some queries are still very slow on strings widget, e.g "*x"
 - same "deadlock" when debug